Move slack calls requiring a user token into the slack adapter helper only used in tests
